Add changes for managing extensions logic
- Add no args constructor to ProductSearcher to comply with case of "no extension plugged"
- Separate logic of searching in the specific webpage, from ProductSearcher to a new class (that would eventually be in the extension)
